+++ lib/grpc/grpc_service_manager.dart
@@ -1,0 +1,49 @@
+import 'package:grpc/grpc_or_grpcweb.dart';
+import 'package:grpc/service_api.dart';
+import 'package:sqlite_wrapper/grpc/authentication_interceptor.dart';
+import 'package:sqlite_wrapper/sqlite_wrapper.dart';
+
+class GrpcServiceManager {
+  late final ClientChannel _channel;
+
+  String? token;
+  // Lazy-initialized service clients
+  AuthClient? _authClient;
+  SqliteWrapperServiceClient? _sqliteClient;
+
+  GrpcServiceManager({
+    String host = 'localhost',
+    int port = 50051,
+    bool secure = false,
+  }) {
+    _channel = GrpcOrGrpcWebClientChannel.toSingleEndpoint(
+      host: host,
+      port: port,
+      transportSecure: secure,
+    );
+  }
+
+  // Getter methods that lazy-initialize clients
+  AuthClient get authClient => _authClient ??= AuthClient(_channel);
+
+  SqliteWrapperServiceClient get sqliteService =>
+      _sqliteClient ??= SqliteWrapperServiceClient(_channel,
+          interceptors: [AuthInterceptor(getToken)]);
+
+  String? getToken() {
+    return token;
+  }
+
+  // Method to add auth token to requests
+  // CallOptions? getAuthOptions(String? token) {
+  //   if (token == null) return null;
+  //   return CallOptions(metadata: {
+  //     'authorization': 'Bearer $token',
+  //   });
+  // }
+
+  // Important: Shutdown the channel when done
+  Future<void> dispose() async {
+    await _channel.shutdown();
+  }
+}