+++ lib/sqlite_wrapper_stub.dart
@@ -5,14 +5,16 @@ import 'sqlite_wrapper_base.dart';
 SQLiteWrapperBase getInstance() =>
     throw UnsupportedError("Unsupported platform");
 
-/// Unused stub implementation just to keep the compiler happy
 class SQLiteWrapperCore extends SQLiteWrapperBase {
   @override
-  Future<DatabaseInfo> openDB(String path,
-      {int version = 0,
-      OnCreate? onCreate,
-      OnUpgrade? onUpgrade,
-      String? dbName}) {
+  Future<DatabaseInfo> openDB(
+    String path, {
+    int version = 0,
+    OnCreate? onCreate,
+    OnUpgrade? onUpgrade,
+    String? dbName,
+    bool useGRPC = false,
+  }) async {
     throw UnimplementedError();
   }
 }